Players Distribution and Constructing par as we go
Players distribution allows for selection of players based off of how often they appear in the real data.

The firstMachines distribution is modified so that it is a cumulative distribution rather than a Monte Carlo one to save memory.

Par is constructed as we go so that sections of prep can be run separately without having to deconstruct par first.